Organize chat functions
